Finishing up credit account issues
