+++ frontend/src/App.tsx
@@ -62,7 +62,8 @@ export const SnackBarContext = createContext({
 });
 
 const App = () => {
-  const queryClient = new QueryClient();
+  // const queryClient = new QueryClient();
+  const [queryClient] = useState(() => new QueryClient());
 
   // Globally used SnackBar component
   const [snackBar, setSnackBar] = useState<SnackBarOptions>({
